Creating a nice conky output for i3bar
In the past I always had problems to create a conky output for i3bar. No
I solved the mystery and was able to create a nice conkyrc file.
Very important things to recognize in future are:

-   To stop flickering, the JSON-Conky must fit in one line (Use
    Backslashes)
-   The seperators are very ugly so I hide them with the method, to give
    them the same color like the background. Now in every JSON-Tag you
    must define a color!
-   Escaping the colors with a Backslash, to not let conky interpret
    them as a Comment.

I also removed the font.conf file, which was created by the font-manager